minimal script for robosuite v140 conversion; created configs
