+++ robomimic/scripts/conversion/convert_to_robosuite_v140_old.py
@@ -1,0 +1,812 @@
+import h5py
+import json
+import argparse
+import os
+from shutil import copyfile
+import robosuite
+import xml.etree.ElementTree as ET
+
+from robosuite.utils.mjcf_utils import find_elements
+
+def replace_elem(parent, old_elem, new_elem):
+    """
+    code adapted from https://stackoverflow.com/a/20931505
+    """
+    parent_index = list(parent).index(old_elem)
+    parent.remove(old_elem)
+    parent.insert(parent_index, new_elem)
+
+def convert_xml(xml_str, env_name):
+    """
+    Postprocess xml string generated by robosuite to be compatible with robosuite v1.3
+    This script should not the xml string if it was already generated using robosuite v1.3
+    Args:
+        xml_str (str): xml string to process (from robosuite v1.2)
+    """
+
+    def get_eef_panda_element(prefix):
+        """
+        Helper function for new element added in robosuite v1.3
+        """
+
+        eef_panda = """
+        <body name="eef" pos="0 0 0.097" quat="1 0 0 0">
+            <site name="grip_site" pos="0 0 0" size="0.01 0.01 0.01" rgba="1 0 0 0.5" type="sphere" group="1"/>
+            <site name="ee_x" pos="0.1 0 0" size="0.005 .1"  quat="0.707105  0 0.707108 0 " rgba="1 0 0 0" type="cylinder" group="1"/>
+            <site name="ee_y" pos="0 0.1 0" size="0.005 .1" quat="0.707105 0.707108 0 0" rgba="0 1 0 0" type="cylinder" group="1"/>
+            <site name="ee_z" pos="0 0 0.1" size="0.005 .1" quat="1 0 0 0" rgba="0 0 1 0" type="cylinder" group="1"/>
+            <!-- This site was added for visualization. -->
+            <site name="grip_site_cylinder" pos="0 0 0" size="0.005 10" rgba="0 1 0 0.3" type="cylinder" group="1"/>
+        </body>
+        """
+
+        tree = ET.ElementTree(ET.fromstring(eef_panda))
+        root = tree.getroot()
+        for elem in root.iter():
+            elem.attrib["name"] = "{}_{}".format(prefix, elem.get("name"))
+
+        return tree
+
+    root = ET.fromstring(xml_str)
+    worldbody = root.find('worldbody')
+    bodies_to_process = []
+    # remove elements that are specific to v1.2
+    for body in worldbody.iter('body'):
+        for site in body.findall('site'):
+            site_name = site.get('name')
+
+            # remove grip_site and grip_site_cylinder
+            if 'grip_site' in site_name:
+                # print("removing {site} from {body}".format(
+                #     site=site_name,
+                #     body=body.get('name'),
+                # ))
+                body.remove(site)
+                if body not in bodies_to_process:
+                    bodies_to_process.append(body)
+
+            # remove site called robot{}_ee{}
+            if site_name.startswith('robot') and 'ee' in site_name:
+                # print("removing {site} from {body}".format(
+                #     site=site_name,
+                #     body=body.get('name'),
+                # ))
+                body.remove(site)
+
+    # add elements introduced in v1.3
+    for body in bodies_to_process:
+        prefix = body.get("name").split('_')[0]
+        eef_tree = get_eef_panda_element(prefix)
+        body.append(eef_tree.getroot())
+        
+        
+    asset = root.find("asset") 
+    for ent in asset.iter("texture"):
+        ent_name = ent.get("name")
+
+        if ent_name in [
+            "Milk_tex-ceramic",
+            "Cereal_tex-cereal",
+            "Can_tex-can",
+        ]:
+            ent.attrib["type"] = "2d"
+
+        # replace can.png -> soda.png for texture path
+        if ent_name == "Can_tex-can":
+            ent_file = ent.get("file")
+            ent_file_split = ent_file.split("/")
+            ent_file_split[-1] = "soda.png"
+            ent.attrib["file"] = '/'.join(ent_file_split)
+
+    for ent in asset.iter("mesh"):
+        ent_name = ent.get("name")
+        if ent_name in [
+            "Milk_milk_mesh",
+            "Cereal_cereal_mesh",
+            "Can_can_mesh",
+        ]:
+            ent_file = ent.get("file")
+            ent_file_split = ent_file.split("/")
+            ent_file_split[-1] = ent_file_split[-1][:-4] + ".msh"
+            ent.attrib["file"] = '/'.join(ent_file_split)
+
+
+    for body in worldbody.iter("body"):
+        body_name = body.get("name")
+        if body_name in ["Milk_main", "Cereal_main", "Can_main"]:
+            inertial = body.find("inertial")
+        else:
+            continue
+
+        if body_name == "Milk_main":
+            inertial.attrib["pos"] = "6.98306e-10 -6.03326e-11 0.00397309"
+            inertial.attrib["mass"] = "0.0191848"
+            inertial.attrib["diaginertia"] = "2.86564e-05 2.83923e-05 4.8063e-06"
+        elif body_name == "Cereal_main":
+            inertial.attrib["pos"] = "2.03178e-07 2.33411e-07 1.28831e-06"
+            inertial.attrib["quat"] = "0.707106 -5.73859e-05 6.76848e-05 0.707107"
+            inertial.attrib["mass"] = "0.0674562"
+            inertial.attrib["diaginertia"] = "0.000182561 0.000131413 6.1265e-05"
+        elif body_name == "Can_main":
+            inertial.attrib["pos"] = "5.89018e-05 -0.000133654 0.0003584"
+            inertial.attrib["quat"] = "0.999769 0.000893507 -3.22198e-05 0.0214535"
+            inertial.attrib["mass"] = "0.0143387"
+            inertial.attrib["diaginertia"] = "9.49915e-06 9.49389e-06 4.38241e-06"
+
+
+
+    # replace robot body
+    robot0_base_str = """
+    <body name="robot0_base" pos="-0.56 0 0.912" gravcomp="0">
+      <inertial pos="0 0 0" mass="0" diaginertia="0 0 0"/>
+      <camera name="robot0_robotview" pos="1 0 0.4" quat="0.653098 0.271041 0.271041 0.653098"/>
+      <body name="robot0_link0" pos="0 0 0" gravcomp="0">
+        <inertial pos="0 0 0.05" mass="4" diaginertia="0.4 0.4 0.4"/>
+        <geom name="robot0_g0_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell010_001" mesh="robot0_link0_vis_0"/>
+        <geom name="robot0_g1_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell009_001" mesh="robot0_link0_vis_1"/>
+        <geom name="robot0_g2_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell003_001" mesh="robot0_link0_vis_2"/>
+        <geom name="robot0_g3_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell_001" mesh="robot0_link0_vis_3"/>
+        <geom name="robot0_g4_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell002_001" mesh="robot0_link0_vis_4"/>
+        <geom name="robot0_g5_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell001_001" mesh="robot0_link0_vis_5"/>
+        <geom name="robot0_g6_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Face636_001" mesh="robot0_link0_vis_6"/>
+        <geom name="robot0_g7_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature018_001" mesh="robot0_link0_vis_7"/>
+        <geom name="robot0_g8_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature019_001" mesh="robot0_link0_vis_8"/>
+        <geom name="robot0_g9_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature023_001" mesh="robot0_link0_vis_9"/>
+        <geom name="robot0_g10_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature022_001" mesh="robot0_link0_vis_10"/>
+        <geom name="robot0_g11_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature017_001" mesh="robot0_link0_vis_11"/>
+        <geom name="robot0_link0_collision" type="mesh" rgba="0 0.5 0 1" mesh="robot0_link0"/>
+        <body name="robot0_link1" pos="0 0 0.333" gravcomp="0">
+          <inertial pos="0 0 -0.07" mass="3" diaginertia="0.3 0.3 0.3"/>
+          <joint name="robot0_joint1" pos="0 0 0" axis="0 0 1" range="-2.8973 2.8973" armature="5" damping="0.1" frictionloss="0.1"/>
+          <geom name="robot0_g12_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature_001" mesh="robot0_link1_vis"/>
+          <geom name="robot0_link1_collision" type="mesh" rgba="0 0.5 0 1" mesh="robot0_link1"/>
+          <body name="robot0_link2" pos="0 0 0" quat="0.707107 -0.707107 0 0" gravcomp="0">
+            <inertial pos="0 -0.1 0" mass="3" diaginertia="0.3 0.3 0.3"/>
+            <joint name="robot0_joint2" pos="0 0 0" axis="0 0 1" range="-1.7628 1.7628" armature="2.5" damping="0.1" frictionloss="0.1"/>
+            <geom name="robot0_g13_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature024" mesh="robot0_link2_vis"/>
+            <geom name="robot0_link2_collision" type="mesh" rgba="0 0.5 0 1" mesh="robot0_link2"/>
+            <body name="robot0_link3" pos="0 -0.316 0" quat="0.707107 0.707107 0 0" gravcomp="0">
+              <inertial pos="0.04 0 -0.05" mass="2" diaginertia="0.2 0.2 0.2"/>
+              <joint name="robot0_joint3" pos="0 0 0" axis="0 0 1" range="-2.8973 2.8973" armature="1.66667" damping="0.1" frictionloss="0.1"/>
+              <geom name="robot0_g14_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature003_004_001_002" mesh="robot0_link3_vis_0"/>
+              <geom name="robot0_g15_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature002_007_001_002" mesh="robot0_link3_vis_1"/>
+              <geom name="robot0_g16_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature001_010_001_002" mesh="robot0_link3_vis_2"/>
+              <geom name="robot0_g17_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature_001_001_001_002" mesh="robot0_link3_vis_3"/>
+              <geom name="robot0_link3_collision" type="mesh" rgba="0 0.5 0 1" mesh="robot0_link3"/>
+              <body name="robot0_link4" pos="0.0825 0 0" quat="0.707107 0.707107 0 0" gravcomp="0">
+                <inertial pos="-0.04 0.05 0" mass="2" diaginertia="0.2 0.2 0.2"/>
+                <joint name="robot0_joint4" pos="0 0 0" axis="0 0 1" range="-3.0718 -0.0698" armature="1.25" damping="0.1" frictionloss="0.1"/>
+                <geom name="robot0_g18_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature001_001_003_001" mesh="robot0_link4_vis_0"/>
+                <geom name="robot0_g19_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature003_001_003_001" mesh="robot0_link4_vis_1"/>
+                <geom name="robot0_g20_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature002_001_003_001" mesh="robot0_link4_vis_2"/>
+                <geom name="robot0_g21_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature_002_003_001" mesh="robot0_link4_vis_3"/>
+                <geom name="robot0_link4_collision" type="mesh" rgba="0 0.5 0 1" mesh="robot0_link4"/>
+                <body name="robot0_link5" pos="-0.0825 0.384 0" quat="0.707107 -0.707107 0 0" gravcomp="0">
+                  <inertial pos="0 0 -0.15" mass="2" diaginertia="0.2 0.2 0.2"/>
+                  <joint name="robot0_joint5" pos="0 0 0" axis="0 0 1" range="-2.8973 2.8973" armature="1" damping="0.1" frictionloss="0.1"/>
+                  <geom name="robot0_g22_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell001_001_001_003" mesh="robot0_link5_vis_0"/>
+                  <geom name="robot0_g23_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell_001_001_003" mesh="robot0_link5_vis_1"/>
+                  <geom name="robot0_g24_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature_002_004_003" mesh="robot0_link5_vis_2"/>
+                  <geom name="robot0_link5_collision" type="mesh" rgba="0 0.5 0 1" mesh="robot0_link5"/>
+                  <body name="robot0_link6" pos="0 0 0" quat="0.707107 0.707107 0 0" gravcomp="0">
+                    <inertial pos="0.06 0 0" mass="1.5" diaginertia="0.1 0.1 0.1"/>
+                    <joint name="robot0_joint6" pos="0 0 0" axis="0 0 1" range="-0.0175 3.7525" armature="0.833333" damping="0.01" frictionloss="0.1"/>
+                    <geom name="robot0_g25_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell006_003_002_001" mesh="robot0_link6_vis_0"/>
+                    <geom name="robot0_g26_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell011_002_002_001" mesh="robot0_link6_vis_1"/>
+                    <geom name="robot0_g27_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell007_002_002_001" mesh="robot0_link6_vis_2"/>
+                    <geom name="robot0_g28_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell005_001_001_002_001" mesh="robot0_link6_vis_3"/>
+                    <geom name="robot0_g29_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell004_001_001_002_001" mesh="robot0_link6_vis_4"/>
+                    <geom name="robot0_g30_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell003_002_001_002_001" mesh="robot0_link6_vis_5"/>
+                    <geom name="robot0_g31_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell002_002_001_002_001" mesh="robot0_link6_vis_6"/>
+                    <geom name="robot0_g32_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Union001_001_001_002_001" mesh="robot0_link6_vis_7"/>
+                    <geom name="robot0_g33_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Union_001_001_002_001" mesh="robot0_link6_vis_8"/>
+                    <geom name="robot0_g34_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Face539_002_001_002_001" mesh="robot0_link6_vis_9"/>
+                    <geom name="robot0_g35_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell_003_001_002_001" mesh="robot0_link6_vis_10"/>
+                    <geom name="robot0_g36_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Face374_002_001_002_001" mesh="robot0_link6_vis_11"/>
+                    <geom name="robot0_g37_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Face065_002_001_002_001" mesh="robot0_link6_vis_12"/>
+                    <geom name="robot0_g38_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Face064_002_001_002_001" mesh="robot0_link6_vis_13"/>
+                    <geom name="robot0_g39_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature002_006_001_002_001" mesh="robot0_link6_vis_14"/>
+                    <geom name="robot0_g40_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Feature001_009_001_002_001" mesh="robot0_link6_vis_15"/>
+                    <geom name="robot0_g41_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Shell012_002_002_001" mesh="robot0_link6_vis_16"/>
+                    <geom name="robot0_link6_collision" type="mesh" rgba="0 0.5 0 1" mesh="robot0_link6"/>
+                    <body name="robot0_link7" pos="0.088 0 0" quat="0.707107 0.707107 0 0" gravcomp="0">
+                      <inertial pos="0 0 0.08" mass="0.5" diaginertia="0.05 0.05 0.05"/>
+                      <joint name="robot0_joint7" pos="0 0 0" axis="0 0 1" range="-2.8973 2.8973" armature="0.714286" damping="0.01" frictionloss="0.1"/>
+                      <geom name="robot0_g42_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Mirroring004_004_002" mesh="robot0_link7_vis_0"/>
+                      <geom name="robot0_g43_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Mirroring001_004_002" mesh="robot0_link7_vis_1"/>
+                      <geom name="robot0_g44_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Mirroring007_004_001" mesh="robot0_link7_vis_2"/>
+                      <geom name="robot0_g45_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Mirroring006_004_001" mesh="robot0_link7_vis_3"/>
+                      <geom name="robot0_g46_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Mirroring005_004_001" mesh="robot0_link7_vis_4"/>
+                      <geom name="robot0_g47_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Mirroring003_004_001" mesh="robot0_link7_vis_5"/>
+                      <geom name="robot0_g48_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Mirroring002_004_001" mesh="robot0_link7_vis_6"/>
+                      <geom name="robot0_g49_vis" type="mesh" contype="0" conaffinity="0" group="1" material="robot0_Part__Mirroring_004_001" mesh="robot0_link7_vis_7"/>
+                      <geom name="robot0_link7_collision" type="mesh" rgba="0 0.5 0 1" mesh="robot0_link7"/>
+                      <body name="robot0_right_hand" pos="0 0 0.1065" quat="0.923785 0 0 -0.382911" gravcomp="0">
+                        <inertial pos="0 0 0" mass="0.5" diaginertia="0.05 0.05 0.05"/>
+                        <camera name="robot0_eye_in_hand" pos="0.05 0 0" quat="0 0.707107 0.707107 0" fovy="75"/>
+                        <body name="gripper0_right_gripper" pos="0 0 0" quat="0.707107 0 0 -0.707107" gravcomp="0">
+                          <inertial pos="0 0 0.17" quat="0.707107 0.707107 0 0" mass="0.3" diaginertia="0.09 0.07 0.05"/>
+                          <geom name="gripper0_hand_visual" quat="0.707107 0 0 0.707107" type="mesh" contype="0" conaffinity="0" group="1" rgba="1 1 1 1" mesh="gripper0_hand_vis"/>
+                          <geom name="gripper0_hand_collision" quat="0.707107 0 0 0.707107" type="mesh" rgba="0 0 0.5 1" mesh="gripper0_hand"/>
+                          <site name="gripper0_ft_frame" pos="0 0 0" size="0.01" group="1" rgba="1 0 0 1"/>
+                          <body name="gripper0_eef" pos="0 0 0.097" gravcomp="0">
+                            <site name="gripper0_grip_site" pos="0 0 0" size="0.01" group="1" rgba="1 0 0 0.5"/>
+                            <site name="gripper0_ee_x" pos="0.1 0 0" quat="0.707105 0 0.707108 0" size="0.005 0.1" group="1" type="cylinder" rgba="1 0 0 0"/>
+                            <site name="gripper0_ee_y" pos="0 0.1 0" quat="0.707105 0.707108 0 0" size="0.005 0.1" group="1" type="cylinder" rgba="0 1 0 0"/>
+                            <site name="gripper0_ee_z" pos="0 0 0.1" size="0.005 0.1" group="1" type="cylinder" rgba="0 0 1 0"/>
+                            <site name="gripper0_grip_site_cylinder" pos="0 0 0" size="0.005 10" group="1" type="cylinder" rgba="0 1 0 0.3"/>
+                          </body>
+                          <body name="gripper0_leftfinger" pos="0 0 0.0524" quat="0.707107 0 0 0.707107" gravcomp="0">
+                            <inertial pos="0 0 0.05" mass="0.1" diaginertia="0.01 0.01 0.005"/>
+                            <joint name="gripper0_finger_joint1" pos="0 0 0" axis="0 1 0" type="slide" range="0 0.04" armature="1" damping="100" frictionloss="1"/>
+                            <geom name="gripper0_finger1_visual" type="mesh" contype="0" conaffinity="0" group="1" rgba="0.499 0.499 0.499 1" mesh="gripper0_finger_vis"/>
+                            <geom name="gripper0_finger1_collision" type="mesh" contype="0" condim="4" rgba="0 0 0.5 1" mesh="gripper0_finger"/>
+                            <body name="gripper0_finger_joint1_tip" pos="0 0.0085 0.056" gravcomp="0">
+                              <inertial pos="0 0 0" quat="0 0 0 1" mass="0.01" diaginertia="0.01 0.01 0.01"/>
+                              <geom name="gripper0_finger1_pad_collision" size="0.008 0.004 0.008" pos="0 -0.005 -0.015" quat="0 0 0 1" type="box" friction="2 0.05 0.0001" solref="0.01 0.5" rgba="0 0 0.5 1"/>
+                            </body>
+                          </body>
+                          <body name="gripper0_rightfinger" pos="0 0 0.0524" quat="0.707107 0 0 0.707107" gravcomp="0">
+                            <inertial pos="0 0 0.05" mass="0.1" diaginertia="0.01 0.01 0.005"/>
+                            <joint name="gripper0_finger_joint2" pos="0 0 0" axis="0 1 0" type="slide" range="-0.04 0" armature="1" damping="100" frictionloss="1"/>
+                            <geom name="gripper0_finger2_visual" quat="0 0 0 1" type="mesh" contype="0" conaffinity="0" group="1" rgba="0.499 0.499 0.499 1" mesh="gripper0_finger_vis"/>
+                            <geom name="gripper0_finger2_collision" quat="0 0 0 1" type="mesh" contype="0" condim="4" rgba="0 0 0.5 1" mesh="gripper0_finger"/>
+                            <body name="gripper0_finger_joint2_tip" pos="0 -0.0085 0.056" gravcomp="0">
+                              <inertial pos="0 0 0" quat="0 0 0 1" mass="0.01" diaginertia="0.01 0.01 0.01"/>
+                              <geom name="gripper0_finger2_pad_collision" size="0.008 0.004 0.008" pos="0 0.005 -0.015" quat="0 0 0 1" type="box" friction="2 0.05 0.0001" solref="0.01 0.5" rgba="0 0 0.5 1"/>
+                            </body>
+                          </body>
+                        </body>
+                      </body>
+                    </body>
+                  </body>
+                </body>
+              </body>
+            </body>
+          </body>
+        </body>
+      </body>
+      <body name="mount0_base" pos="0 0 0.01" gravcomp="0">
+        <body name="mount0_controller_box" pos="0 0 0" gravcomp="0">
+          <inertial pos="-0.325 0 -0.38" mass="46.64" diaginertia="1.71363 1.27988 0.809981"/>
+          <geom name="mount0_controller_box_col" size="0.11 0.2 0.265" pos="-0.325 0 -0.38" type="box" rgba="0.5 0.5 0 1"/>
+        </body>
+        <body name="mount0_pedestal_feet" pos="0 0 0" gravcomp="0">
+          <inertial pos="-0.1225 0 -0.758" mass="167.09" diaginertia="8.16095 9.59375 15.0785"/>
+          <geom name="mount0_pedestal_feet_col" size="0.385 0.35 0.155" pos="-0.1225 0 -0.758" type="box" rgba="0.5 0.5 0 1"/>
+        </body>
+        <body name="mount0_torso" pos="0 0 0" gravcomp="0">
+          <inertial pos="0 0 0" mass="0.0001" diaginertia="1e-08 1e-08 1e-08"/>
+          <geom name="mount0_torso_vis" size="0.05 0.05 0.05" pos="0 0 -0.05" type="box" contype="0" conaffinity="0" group="1" rgba="0.2 0.2 0.2 1"/>
+        </body>
+        <body name="mount0_pedestal" pos="0 0 0" gravcomp="0">
+          <inertial pos="0 0 0" quat="0.659267 -0.259505 -0.260945 0.655692" mass="60.864" diaginertia="6.0869 5.81635 4.20915"/>
+          <geom name="mount0_pedestal_vis" type="mesh" contype="0" conaffinity="0" group="1" rgba="0.2 0.2 0.2 1" mesh="mount0_pedestal"/>
+          <geom name="mount0_pedestal_col" size="0.18 0.31" pos="-0.02 0 -0.29" type="cylinder" rgba="0.5 0.5 0 1"/>
+        </body>
+      </body>
+    </body>
+    """
+
+    
+    textures = """
+    <texture type="skybox" builtin="gradient" rgb1="0.9 0.9 1" rgb2="0.2 0.3 0.4" width="256" height="1536"/>
+    <texture type="2d" name="texplane" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/light-gray-floor-tile.png"/>
+    """
+
+    if env_name == "Lift":
+      textures += """
+      <texture type="cube" name="tex-ceramic" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/ceramic.png"/>
+      <texture type="cube" name="tex-steel-brushed" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/steel-brushed.png"/>
+      <texture type="2d" name="tex-light-gray-plaster" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/light-gray-plaster.png"/>
+      <texture type="cube" name="textable" builtin="flat" rgb1="0.5 0.5 0.5" width="512" height="3072"/>
+      <texture type="cube" name="cube_redwood" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/textures/red-wood.png"/>
+      """
+    elif env_name == "PickPlaceCan":
+      textures += """
+      <texture type="cube" name="tex-steel-brushed" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/steel-brushed.png"/>
+      <texture type="2d" name="tex-light-gray-plaster" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/light-gray-plaster.png"/>
+      <texture type="2d" name="tex-light-wood" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/light-wood.png"/>
+      <texture type="2d" name="tex-dark-wood" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/dark-wood.png"/>
+      <texture type="2d" name="Milk_tex-ceramic" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/objects/../textures/ceramic.png"/>
+      <texture type="2d" name="Bread_tex-bread" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/objects/../textures/bread.png"/>
+      <texture type="2d" name="Cereal_tex-cereal" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/objects/../textures/cereal.png"/>
+      <texture type="2d" name="Can_tex-can" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/objects/../textures/soda.png"/>
+      """
+    elif env_name == "NutAssemblySquare":
+      textures += """
+      <texture type="cube" name="steel-metal" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/steel-scratched.png"/>
+      <texture type="cube" name="brass-metal" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/brass-ambra.png"/>
+      <texture type="cube" name="tex-ceramic" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/ceramic.png"/>
+      <texture type="cube" name="tex-steel-brushed" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/steel-brushed.png"/>
+      <texture type="2d" name="tex-light-gray-plaster" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/light-gray-plaster.png"/>
+      <texture type="cube" name="textable" builtin="flat" rgb1="0.5 0.5 0.5" width="512" height="3072"/>
+      <texture type="cube" name="SquareNut_brass-metal" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/objects/../textures/brass-ambra.png"/>
+      <texture type="cube" name="RoundNut_steel-metal" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/objects/../textures/steel-scratched.png"/>
+      """
+    elif env_name == "ToolHang":
+      textures += """
+      <texture type="cube" name="tex-ceramic" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/ceramic.png"/>
+      <texture type="cube" name="tex-steel-brushed" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/steel-brushed.png"/>
+      <texture type="2d" name="tex-light-gray-plaster" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/light-gray-plaster.png"/>
+      <texture type="cube" name="textable" builtin="flat" rgb1="0.5 0.5 0.5" width="512" height="3072"/>
+      <texture type="cube" name="stand_brass" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/textures/brass-ambra.png"/>
+      <texture type="cube" name="frame_brass" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/textures/brass-ambra.png"/>
+      <texture type="cube" name="frame_ceramic" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/textures/ceramic.png"/>
+      <texture type="cube" name="frame_steel" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/textures/steel-scratched.png"/>
+      <texture type="cube" name="tool_hole1_steel" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/textures/steel-scratched.png"/>
+      <texture type="cube" name="tool_hole2_steel" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/textures/steel-scratched.png"/>
+      <texture type="cube" name="tool_handle_steel" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/textures/steel-scratched.png"/>
+      """
+    elif env_name == "TwoArmTransport":
+      textures += """
+      <texture type="cube" name="tex-ceramic" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/ceramic.png"/>
+      <texture type="cube" name="tex-steel-brushed" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/steel-brushed.png"/>
+      <texture type="2d" name="tex-light-gray-plaster" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/arenas/../textures/light-gray-plaster.png"/>
+      <texture type="cube" name="textable" builtin="flat" rgb1="0.5 0.5 0.5" width="512" height="3072"/>
+      <texture type="cube" name="payload_metal" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/textures/steel-scratched.png"/>
+      <texture type="cube" name="payload_wood" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/textures/light-wood.png"/>
+      <texture type="cube" name="trash_redwood" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/textures/red-wood.png"/>
+      <texture type="cube" name="transport_start_bin_dark_wood" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/textures/dark-wood.png"/>
+      <texture type="cube" name="transport_target_bin_dark_wood" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/textures/dark-wood.png"/>
+      <texture type="cube" name="transport_trash_bin_dark_wood" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/textures/dark-wood.png"/>
+      <texture type="cube" name="transport_start_bin_lid_dark_wood" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/textures/dark-wood.png"/>
+      """
+    else:
+      raise NotImplementedError
+
+    materials = """
+    <material name="robot0_Face636_001" shininess="0.45" rgba="0.901961 0.921569 0.929412 1"/>
+    <material name="robot0_Part__Feature017_001" shininess="0.45"/>
+    <material name="robot0_Part__Feature018_001" shininess="0.45"/>
+    <material name="robot0_Part__Feature019_001" shininess="0.45"/>
+    <material name="robot0_Part__Feature022_001" shininess="0.45" rgba="0.901961 0.921569 0.929412 1"/>
+    <material name="robot0_Part__Feature023_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Shell001_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Shell002_001" shininess="0.45" rgba="0.901961 0.921569 0.929412 1"/>
+    <material name="robot0_Shell003_001" shininess="0.45" rgba="0.901961 0.921569 0.929412 1"/>
+    <material name="robot0_Shell009_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Shell010_001" shininess="0.45" rgba="0.901961 0.921569 0.929412 1"/>
+    <material name="robot0_Shell_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Part__Feature_001" shininess="0.45"/>
+    <material name="robot0_Part__Feature024" shininess="0.45"/>
+    <material name="robot0_Part__Feature001_010_001_002" shininess="0.45"/>
+    <material name="robot0_Part__Feature002_007_001_002" shininess="0.45"/>
+    <material name="robot0_Part__Feature003_004_001_002" shininess="0.45"/>
+    <material name="robot0_Part__Feature_001_001_001_002" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Part__Feature001_001_003_001" shininess="0.45"/>
+    <material name="robot0_Part__Feature002_001_003_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Part__Feature003_001_003_001" shininess="0.45"/>
+    <material name="robot0_Part__Feature_002_003_001" shininess="0.45"/>
+    <material name="robot0_Part__Feature_002_004_003" shininess="0.45"/>
+    <material name="robot0_Shell001_001_001_003" shininess="0.45" rgba="0.25 0.25 0.25 1"/>
+    <material name="robot0_Shell_001_001_003" shininess="0.45"/>
+    <material name="robot0_Face064_002_001_002_001" shininess="0.45" rgba="1 0 0 1"/>
+    <material name="robot0_Face065_002_001_002_001" shininess="0.45" rgba="0 1 0 1"/>
+    <material name="robot0_Face374_002_001_002_001" shininess="0.45"/>
+    <material name="robot0_Face539_002_001_002_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Part__Feature001_009_001_002_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Part__Feature002_006_001_002_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Shell002_002_001_002_001" shininess="0.45"/>
+    <material name="robot0_Shell003_002_001_002_001" shininess="0.45"/>
+    <material name="robot0_Shell004_001_001_002_001" shininess="0.45"/>
+    <material name="robot0_Shell005_001_001_002_001" shininess="0.45"/>
+    <material name="robot0_Shell006_003_002_001" shininess="0.45" rgba="0.901961 0.921569 0.929412 1"/>
+    <material name="robot0_Shell007_002_002_001" shininess="0.45" rgba="0.25 0.25 0.25 1"/>
+    <material name="robot0_Shell011_002_002_001" shininess="0.45"/>
+    <material name="robot0_Shell012_002_002_001" shininess="0.45"/>
+    <material name="robot0_Shell_003_001_002_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Union001_001_001_002_001" shininess="0.45" rgba="0.039216 0.541176 0.780392 1"/>
+    <material name="robot0_Union_001_001_002_001" shininess="0.45" rgba="0.039216 0.541176 0.780392 1"/>
+    <material name="robot0_Part__Mirroring001_004_002" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Part__Mirroring002_004_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Part__Mirroring003_004_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Part__Mirroring004_004_002" shininess="0.45"/>
+    <material name="robot0_Part__Mirroring005_004_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Part__Mirroring006_004_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Part__Mirroring007_004_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+    <material name="robot0_Part__Mirroring_004_001" shininess="0.45" rgba="0.898039 0.917647 0.929412 1"/>"""
+    
+    if env_name == "PickPlaceCan":
+      materials = """
+      <material name="floorplane" texture="texplane" texuniform="true" texrepeat="2 2" specular="0" shininess="0" reflectance="0.01"/>
+      <material name="table_legs_metal" texture="tex-steel-brushed" shininess="0.8" reflectance="0.8"/>
+      <material name="walls_mat" texture="tex-light-gray-plaster" texuniform="true" texrepeat="3 3" specular="0.1" shininess="0.1"/>
+      <material name="light-wood" texture="tex-light-wood" texuniform="true" texrepeat="15 15"/>
+      <material name="dark-wood" texture="tex-dark-wood" texuniform="true" texrepeat="5 5"/>
+      """ + materials
+      
+      materials += """
+      <material name="Milk_ceramic" texture="Milk_tex-ceramic" texuniform="true" reflectance="0.5"/>
+      <material name="Bread_bread" texture="Bread_tex-bread" texuniform="true" texrepeat="15 15" reflectance="0.7"/>
+      <material name="Cereal_cereal" texture="Cereal_tex-cereal" reflectance="0.5"/>
+      <material name="Can_coke" texture="Can_tex-can" texuniform="true" texrepeat="5 5" reflectance="0.7"/>
+      """
+    elif env_name == "NutAssemblySquare":
+      materials = """
+      <material name="floorplane" texture="texplane" texuniform="true" texrepeat="2 2" specular="0" shininess="0" reflectance="0.01"/>
+      <material name="smetal" texture="steel-metal" texuniform="true" specular="1" shininess="1" reflectance="1"/>
+      <material name="bmetal" texture="brass-metal" texuniform="true" specular="1" shininess="1" reflectance="1"/>
+      <material name="table_ceramic" texture="tex-ceramic" specular="0.2" shininess="0"/>
+      <material name="table_legs_metal" texture="tex-steel-brushed" shininess="0.8" reflectance="0.8"/>
+      <material name="walls_mat" texture="tex-light-gray-plaster" texuniform="true" texrepeat="3 3" specular="0.1" shininess="0.1"/>
+      <material name="table_mat" texture="textable"/>
+      """ + materials
+
+      materials += """
+      <material name="SquareNut_bmetal" texture="SquareNut_brass-metal" texuniform="true" specular="1" shininess="1" reflectance="1"/>
+      <material name="RoundNut_smetal" texture="RoundNut_steel-metal" texuniform="true" specular="1" shininess="1" reflectance="1"/>
+      """
+    elif env_name == "Lift":
+      materials = """
+      <material name="floorplane" texture="texplane" texuniform="true" texrepeat="2 2" specular="0" shininess="0" reflectance="0.01"/>
+      <material name="table_ceramic" texture="tex-ceramic" specular="0.2" shininess="0"/>
+      <material name="table_legs_metal" texture="tex-steel-brushed" shininess="0.8" reflectance="0.8"/>
+      <material name="walls_mat" texture="tex-light-gray-plaster" texuniform="true" texrepeat="3 3" specular="0.1" shininess="0.1"/>
+      <material name="table_mat" texture="textable"/>
+      """ + materials
+
+      materials += """
+      <material name="cube_redwood_mat" texture="cube_redwood" specular="0.4" shininess="0.1"/>
+      """
+    elif env_name == "ToolHang":
+      materials = """
+      <material name="floorplane" texture="texplane" texuniform="true" texrepeat="2 2" specular="0" shininess="0" reflectance="0.01"/>
+      <material name="table_ceramic" texture="tex-ceramic" specular="0.2" shininess="0"/>
+      <material name="table_legs_metal" texture="tex-steel-brushed" shininess="0.8" reflectance="0.8"/>
+      <material name="walls_mat" texture="tex-light-gray-plaster" texuniform="true" texrepeat="3 3" specular="0.1" shininess="0.1"/>
+      <material name="table_mat" texture="textable"/>
+      """ + materials
+
+      materials += """
+      <material name="stand_brass_mat" texture="stand_brass" texrepeat="3 3" specular="0.4" shininess="0.1"/>
+      <material name="frame_brass_mat" texture="frame_brass" texrepeat="3 3" specular="0.4" shininess="0.1"/>
+      <material name="frame_ceramic_mat" texture="frame_ceramic" texrepeat="3 3" specular="0.4" shininess="0.1"/>
+      <material name="frame_steel_mat" texture="frame_steel" texrepeat="3 3" specular="0.4" shininess="0.1"/>
+      <material name="tool_hole1_steel_mat" texture="tool_hole1_steel" texrepeat="3 3" specular="0.4" shininess="0.1"/>
+      <material name="tool_hole2_steel_mat" texture="tool_hole2_steel" texrepeat="3 3" specular="0.4" shininess="0.1"/>
+      <material name="tool_handle_steel_mat" texture="tool_handle_steel" texrepeat="3 3" specular="0.4" shininess="0.1"/>
+      """
+    elif env_name == "TwoArmTransport":
+      materials = """
+      <material name="floorplane" texture="texplane" texuniform="true" texrepeat="2 2" specular="0" shininess="0" reflectance="0.01"/>
+      <material name="table_ceramic" texture="tex-ceramic" specular="0.2" shininess="0"/>
+      <material name="table_legs_metal" texture="tex-steel-brushed" shininess="0.8" reflectance="0.8"/>
+      <material name="walls_mat" texture="tex-light-gray-plaster" texuniform="true" texrepeat="3 3" specular="0.1" shininess="0.1"/>
+      <material name="table_mat" texture="textable"/>
+      """ + materials
+
+      materials += """
+      <material name="robot1_Face636_001" shininess="0.45" rgba="0.901961 0.921569 0.929412 1"/>
+      <material name="robot1_Part__Feature017_001" shininess="0.45"/>
+      <material name="robot1_Part__Feature018_001" shininess="0.45"/>
+      <material name="robot1_Part__Feature019_001" shininess="0.45"/>
+      <material name="robot1_Part__Feature022_001" shininess="0.45" rgba="0.901961 0.921569 0.929412 1"/>
+      <material name="robot1_Part__Feature023_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Shell001_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Shell002_001" shininess="0.45" rgba="0.901961 0.921569 0.929412 1"/>
+      <material name="robot1_Shell003_001" shininess="0.45" rgba="0.901961 0.921569 0.929412 1"/>
+      <material name="robot1_Shell009_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Shell010_001" shininess="0.45" rgba="0.901961 0.921569 0.929412 1"/>
+      <material name="robot1_Shell_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Part__Feature_001" shininess="0.45"/>
+      <material name="robot1_Part__Feature024" shininess="0.45"/>
+      <material name="robot1_Part__Feature001_010_001_002" shininess="0.45"/>
+      <material name="robot1_Part__Feature002_007_001_002" shininess="0.45"/>
+      <material name="robot1_Part__Feature003_004_001_002" shininess="0.45"/>
+      <material name="robot1_Part__Feature_001_001_001_002" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Part__Feature001_001_003_001" shininess="0.45"/>
+      <material name="robot1_Part__Feature002_001_003_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Part__Feature003_001_003_001" shininess="0.45"/>
+      <material name="robot1_Part__Feature_002_003_001" shininess="0.45"/>
+      <material name="robot1_Part__Feature_002_004_003" shininess="0.45"/>
+      <material name="robot1_Shell001_001_001_003" shininess="0.45" rgba="0.25 0.25 0.25 1"/>
+      <material name="robot1_Shell_001_001_003" shininess="0.45"/>
+      <material name="robot1_Face064_002_001_002_001" shininess="0.45" rgba="1 0 0 1"/>
+      <material name="robot1_Face065_002_001_002_001" shininess="0.45" rgba="0 1 0 1"/>
+      <material name="robot1_Face374_002_001_002_001" shininess="0.45"/>
+      <material name="robot1_Face539_002_001_002_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Part__Feature001_009_001_002_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Part__Feature002_006_001_002_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Shell002_002_001_002_001" shininess="0.45"/>
+      <material name="robot1_Shell003_002_001_002_001" shininess="0.45"/>
+      <material name="robot1_Shell004_001_001_002_001" shininess="0.45"/>
+      <material name="robot1_Shell005_001_001_002_001" shininess="0.45"/>
+      <material name="robot1_Shell006_003_002_001" shininess="0.45" rgba="0.901961 0.921569 0.929412 1"/>
+      <material name="robot1_Shell007_002_002_001" shininess="0.45" rgba="0.25 0.25 0.25 1"/>
+      <material name="robot1_Shell011_002_002_001" shininess="0.45"/>
+      <material name="robot1_Shell012_002_002_001" shininess="0.45"/>
+      <material name="robot1_Shell_003_001_002_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Union001_001_001_002_001" shininess="0.45" rgba="0.039216 0.541176 0.780392 1"/>
+      <material name="robot1_Union_001_001_002_001" shininess="0.45" rgba="0.039216 0.541176 0.780392 1"/>
+      <material name="robot1_Part__Mirroring001_004_002" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Part__Mirroring002_004_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Part__Mirroring003_004_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Part__Mirroring004_004_002" shininess="0.45"/>
+      <material name="robot1_Part__Mirroring005_004_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Part__Mirroring006_004_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Part__Mirroring007_004_001" shininess="0.45" rgba="0.25098 0.25098 0.25098 1"/>
+      <material name="robot1_Part__Mirroring_004_001" shininess="0.45" rgba="0.898039 0.917647 0.929412 1"/>
+      <material name="payload_metal_mat" texture="payload_metal" texrepeat="3 3" specular="0.4" shininess="0.1"/>
+      <material name="payload_wood_mat" texture="payload_wood" texrepeat="3 3" specular="0.4" shininess="0.1"/>
+      <material name="trash_redwood_mat" texture="trash_redwood" texrepeat="3 3" specular="0.4" shininess="0.1"/>
+      <material name="transport_start_bin_dark_wood_mat" texture="transport_start_bin_dark_wood" texrepeat="3 3" specular="0.4" shininess="0.1"/>
+      <material name="transport_target_bin_dark_wood_mat" texture="transport_target_bin_dark_wood" texrepeat="3 3" specular="0.4" shininess="0.1"/>
+      <material name="transport_trash_bin_dark_wood_mat" texture="transport_trash_bin_dark_wood" texrepeat="3 3" specular="0.4" shininess="0.1"/>
+      <material name="transport_start_bin_lid_dark_wood_mat" texture="transport_start_bin_lid_dark_wood" texrepeat="3 3" specular="0.4" shininess="0.1"/>
+      """
+    else:
+      raise NotImplementedError
+
+    meshes = """
+    <mesh name="robot0_link0" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link0.stl"/>
+    <mesh name="robot0_link1" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link1.stl"/>
+    <mesh name="robot0_link2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link2.stl"/>
+    <mesh name="robot0_link3" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link3.stl"/>
+    <mesh name="robot0_link4" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link4.stl"/>
+    <mesh name="robot0_link5" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link5.stl"/>
+    <mesh name="robot0_link6" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link6.stl"/>
+    <mesh name="robot0_link7" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link7.stl"/>
+    <mesh name="robot0_link0_vis_0" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_0.obj"/>
+    <mesh name="robot0_link0_vis_1" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_1.obj"/>
+    <mesh name="robot0_link0_vis_2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_2.obj"/>
+    <mesh name="robot0_link0_vis_3" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_3.obj"/>
+    <mesh name="robot0_link0_vis_4" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_4.obj"/>
+    <mesh name="robot0_link0_vis_5" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_5.obj"/>
+    <mesh name="robot0_link0_vis_6" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_6.obj"/>
+    <mesh name="robot0_link0_vis_7" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_7.obj"/>
+    <mesh name="robot0_link0_vis_8" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_8.obj"/>
+    <mesh name="robot0_link0_vis_9" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_9.obj"/>
+    <mesh name="robot0_link0_vis_10" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_10.obj"/>
+    <mesh name="robot0_link0_vis_11" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_11.obj"/>
+    <mesh name="robot0_link1_vis" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link1_vis/link1_vis.obj"/>
+    <mesh name="robot0_link2_vis" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link2_vis/link2_vis.obj"/>
+    <mesh name="robot0_link3_vis_0" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link3_vis/link3_vis_0.obj"/>
+    <mesh name="robot0_link3_vis_1" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link3_vis/link3_vis_1.obj"/>
+    <mesh name="robot0_link3_vis_2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link3_vis/link3_vis_2.obj"/>
+    <mesh name="robot0_link3_vis_3" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link3_vis/link3_vis_3.obj"/>
+    <mesh name="robot0_link4_vis_0" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link4_vis/link4_vis_0.obj"/>
+    <mesh name="robot0_link4_vis_1" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link4_vis/link4_vis_1.obj"/>
+    <mesh name="robot0_link4_vis_2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link4_vis/link4_vis_2.obj"/>
+    <mesh name="robot0_link4_vis_3" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link4_vis/link4_vis_3.obj"/>
+    <mesh name="robot0_link5_vis_0" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link5_vis/link5_vis_0.obj"/>
+    <mesh name="robot0_link5_vis_1" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link5_vis/link5_vis_1.obj"/>
+    <mesh name="robot0_link5_vis_2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link5_vis/link5_vis_2.obj"/>
+    <mesh name="robot0_link6_vis_0" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_0.obj"/>
+    <mesh name="robot0_link6_vis_1" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_1.obj"/>
+    <mesh name="robot0_link6_vis_2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_2.obj"/>
+    <mesh name="robot0_link6_vis_3" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_3.obj"/>
+    <mesh name="robot0_link6_vis_4" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_4.obj"/>
+    <mesh name="robot0_link6_vis_5" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_5.obj"/>
+    <mesh name="robot0_link6_vis_6" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_6.obj"/>
+    <mesh name="robot0_link6_vis_7" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_7.obj"/>
+    <mesh name="robot0_link6_vis_8" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_8.obj"/>
+    <mesh name="robot0_link6_vis_9" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_9.obj"/>
+    <mesh name="robot0_link6_vis_10" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_10.obj"/>
+    <mesh name="robot0_link6_vis_11" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_11.obj"/>
+    <mesh name="robot0_link6_vis_12" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_12.obj"/>
+    <mesh name="robot0_link6_vis_13" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_13.obj"/>
+    <mesh name="robot0_link6_vis_14" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_14.obj"/>
+    <mesh name="robot0_link6_vis_15" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_15.obj"/>
+    <mesh name="robot0_link6_vis_16" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_16.obj"/>
+    <mesh name="robot0_link7_vis_0" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_0.obj"/>
+    <mesh name="robot0_link7_vis_1" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_1.obj"/>
+    <mesh name="robot0_link7_vis_2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_2.obj"/>
+    <mesh name="robot0_link7_vis_3" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_3.obj"/>
+    <mesh name="robot0_link7_vis_4" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_4.obj"/>
+    <mesh name="robot0_link7_vis_5" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_5.obj"/>
+    <mesh name="robot0_link7_vis_6" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_6.obj"/>
+    <mesh name="robot0_link7_vis_7" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_7.obj"/>
+    <mesh name="mount0_pedestal" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/mounts/meshes/rethink_mount/pedestal.stl"/>
+    <mesh name="gripper0_hand" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/grippers/meshes/panda_gripper/hand.stl"/>
+    <mesh name="gripper0_hand_vis" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/grippers/meshes/panda_gripper/hand_vis.stl"/>
+    <mesh name="gripper0_finger" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/grippers/meshes/panda_gripper/finger.stl"/>
+    <mesh name="gripper0_finger_vis" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/grippers/meshes/panda_gripper/finger_vis.stl"/>
+    <mesh name="gripper0_finger_vis2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/grippers/meshes/panda_gripper/finger_longer.stl"/>"""
+
+
+    if env_name == "PickPlaceCan":
+      meshes += """
+      <mesh name="VisualMilk_milk_mesh" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/objects/meshes/milk.stl" scale="0.9 0.9 0.9"/>
+      <mesh name="VisualBread_bread_mesh" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/objects/meshes/bread.stl" scale="0.8 0.8 0.8"/>
+      <mesh name="VisualCereal_cereal_mesh" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/objects/meshes/cereal.stl"/>
+      <mesh name="VisualCan_can_mesh" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/objects/meshes/can.stl"/>
+      <mesh name="Milk_milk_mesh" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/objects/meshes/milk.msh" scale="0.9 0.9 0.9"/>
+      <mesh name="Bread_bread_mesh" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/objects/meshes/bread.stl" scale="0.8 0.8 0.8"/>
+      <mesh name="Cereal_cereal_mesh" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/objects/meshes/cereal.msh"/>
+      <mesh name="Can_can_mesh" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/objects/meshes/can.msh"/>
+      """
+    elif env_name == "TwoArmTransport":
+      meshes += """
+      <mesh name="robot1_link0" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link0.stl"/>
+      <mesh name="robot1_link1" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link1.stl"/>
+      <mesh name="robot1_link2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link2.stl"/>
+      <mesh name="robot1_link3" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link3.stl"/>
+      <mesh name="robot1_link4" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link4.stl"/>
+      <mesh name="robot1_link5" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link5.stl"/>
+      <mesh name="robot1_link6" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link6.stl"/>
+      <mesh name="robot1_link7" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/meshes/link7.stl"/>
+      <mesh name="robot1_link0_vis_0" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_0.obj"/>
+      <mesh name="robot1_link0_vis_1" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_1.obj"/>
+      <mesh name="robot1_link0_vis_2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_2.obj"/>
+      <mesh name="robot1_link0_vis_3" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_3.obj"/>
+      <mesh name="robot1_link0_vis_4" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_4.obj"/>
+      <mesh name="robot1_link0_vis_5" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_5.obj"/>
+      <mesh name="robot1_link0_vis_6" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_6.obj"/>
+      <mesh name="robot1_link0_vis_7" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_7.obj"/>
+      <mesh name="robot1_link0_vis_8" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_8.obj"/>
+      <mesh name="robot1_link0_vis_9" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_9.obj"/>
+      <mesh name="robot1_link0_vis_10" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_10.obj"/>
+      <mesh name="robot1_link0_vis_11" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link0_vis/link0_vis_11.obj"/>
+      <mesh name="robot1_link1_vis" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link1_vis/link1_vis.obj"/>
+      <mesh name="robot1_link2_vis" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link2_vis/link2_vis.obj"/>
+      <mesh name="robot1_link3_vis_0" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link3_vis/link3_vis_0.obj"/>
+      <mesh name="robot1_link3_vis_1" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link3_vis/link3_vis_1.obj"/>
+      <mesh name="robot1_link3_vis_2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link3_vis/link3_vis_2.obj"/>
+      <mesh name="robot1_link3_vis_3" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link3_vis/link3_vis_3.obj"/>
+      <mesh name="robot1_link4_vis_0" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link4_vis/link4_vis_0.obj"/>
+      <mesh name="robot1_link4_vis_1" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link4_vis/link4_vis_1.obj"/>
+      <mesh name="robot1_link4_vis_2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link4_vis/link4_vis_2.obj"/>
+      <mesh name="robot1_link4_vis_3" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link4_vis/link4_vis_3.obj"/>
+      <mesh name="robot1_link5_vis_0" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link5_vis/link5_vis_0.obj"/>
+      <mesh name="robot1_link5_vis_1" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link5_vis/link5_vis_1.obj"/>
+      <mesh name="robot1_link5_vis_2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link5_vis/link5_vis_2.obj"/>
+      <mesh name="robot1_link6_vis_0" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_0.obj"/>
+      <mesh name="robot1_link6_vis_1" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_1.obj"/>
+      <mesh name="robot1_link6_vis_2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_2.obj"/>
+      <mesh name="robot1_link6_vis_3" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_3.obj"/>
+      <mesh name="robot1_link6_vis_4" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_4.obj"/>
+      <mesh name="robot1_link6_vis_5" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_5.obj"/>
+      <mesh name="robot1_link6_vis_6" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_6.obj"/>
+      <mesh name="robot1_link6_vis_7" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_7.obj"/>
+      <mesh name="robot1_link6_vis_8" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_8.obj"/>
+      <mesh name="robot1_link6_vis_9" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_9.obj"/>
+      <mesh name="robot1_link6_vis_10" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_10.obj"/>
+      <mesh name="robot1_link6_vis_11" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_11.obj"/>
+      <mesh name="robot1_link6_vis_12" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_12.obj"/>
+      <mesh name="robot1_link6_vis_13" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_13.obj"/>
+      <mesh name="robot1_link6_vis_14" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_14.obj"/>
+      <mesh name="robot1_link6_vis_15" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_15.obj"/>
+      <mesh name="robot1_link6_vis_16" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link6_vis/link6_vis_16.obj"/>
+      <mesh name="robot1_link7_vis_0" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_0.obj"/>
+      <mesh name="robot1_link7_vis_1" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_1.obj"/>
+      <mesh name="robot1_link7_vis_2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_2.obj"/>
+      <mesh name="robot1_link7_vis_3" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_3.obj"/>
+      <mesh name="robot1_link7_vis_4" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_4.obj"/>
+      <mesh name="robot1_link7_vis_5" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_5.obj"/>
+      <mesh name="robot1_link7_vis_6" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_6.obj"/>
+      <mesh name="robot1_link7_vis_7" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/robots/panda/obj_meshes/link7_vis/link7_vis_7.obj"/>
+      <mesh name="mount1_pedestal" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/mounts/meshes/rethink_mount/pedestal.stl"/>
+      <mesh name="gripper1_hand" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/grippers/meshes/panda_gripper/hand.stl"/>
+      <mesh name="gripper1_hand_vis" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/grippers/meshes/panda_gripper/hand_vis.stl"/>
+      <mesh name="gripper1_finger" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/grippers/meshes/panda_gripper/finger.stl"/>
+      <mesh name="gripper1_finger_vis" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/grippers/meshes/panda_gripper/finger_vis.stl"/>
+      <mesh name="gripper1_finger_vis2" file="/home/soroushn/code/robosuite-dev/robosuite/models/assets/grippers/meshes/panda_gripper/finger_longer.stl"/>
+      """
+    elif env_name in ["NutAssemblySquare", "Lift", "ToolHang"]:
+      # no meshes to add
+      pass
+    else:
+      raise NotImplementedError
+    
+    asset_str = """<asset>{textures}{materials}{meshes}\n</asset>""".format(
+      textures=textures, materials=materials, meshes=meshes
+    )
+
+    # print(asset_str)
+    # exit()
+
+    asset_new = ET.ElementTree(ET.fromstring(asset_str)).getroot()
+    asset_old = find_elements(
+        root=root,
+        tags="asset",
+        return_first=True
+    )
+    replace_elem(root, asset_old, asset_new)
+
+    robot0_base_new = ET.ElementTree(ET.fromstring(robot0_base_str)).getroot()
+    if env_name == "TwoArmTransport":
+      robot0_base_new.set("pos", "0 -0.81 0.912")
+      robot0_base_new.set("quat", "0.707107 0 0 0.707107")
+    elif env_name == "PickPlaceCan":
+      robot0_base_new.set("pos", "-0.5 -0.1 0.912")
+    robot0_base_old = find_elements(
+        root=worldbody,
+        tags="body",
+        attribs={"name": "robot0_base"},
+        return_first=True
+    )
+    replace_elem(worldbody, robot0_base_old, robot0_base_new)
+
+    if env_name == "TwoArmTransport":
+      robot1_base_str = robot0_base_str.replace("robot0", "robot1")
+      robot1_base_str = robot1_base_str.replace("gripper0", "gripper1")
+      robot1_base_str = robot1_base_str.replace("mount0", "mount1")
+      robot1_base_new = ET.ElementTree(ET.fromstring(robot1_base_str)).getroot()
+      robot1_base_new.set("pos", "0 0.81 0.912")
+      robot1_base_new.set("quat", "0.707107 0 0 -0.707107")
+      robot1_base_old = find_elements(
+          root=worldbody,
+          tags="body",
+          attribs={"name": "robot1_base"},
+          return_first=True
+      )
+      replace_elem(worldbody, robot1_base_old, robot1_base_new)
+
+
+    compiler = find_elements(
+        root=root,
+        tags="compiler",
+        return_first=True
+    )
+    compiler.set("autolimits", "true")
+
+    floor = find_elements(
+        root=worldbody,
+        tags="geom",
+        attribs={"name": "floor"},
+        return_first=True
+    )
+    floor.set("group", "1")
+
+    if env_name == "Lift":
+      light_str = """
+      <light pos="1 1 1.5" dir="-0.19245 -0.19245 -0.96225" directional="true" castshadow="false"/>
+      """
+      light_new = ET.ElementTree(ET.fromstring(light_str)).getroot()
+      light_old = find_elements(
+          root=worldbody,
+          tags="light",
+          attribs={"name": "light1"},
+          return_first=True
+      )
+      replace_elem(worldbody, light_old, light_new)
+      worldbody.remove(find_elements(
+          root=worldbody,
+          tags="light",
+          attribs={"name": "light2"},
+          return_first=True
+      ))
+
+    xml_str = ET.tostring(root, encoding="utf8").decode("utf8")
+    return xml_str
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "--dataset",
+        type=str,
+        help="path to input hdf5 dataset",
+    )
+    parser.add_argument(
+        "--output_dataset",
+        type=str,
+        help="path to output hdf5 dataset",
+    )
+    args = parser.parse_args()
+    
+    args.dataset = os.path.expanduser(args.dataset)
+    args.output_dataset = os.path.expanduser(args.output_dataset)
+    
+    assert args.output_dataset != args.dataset
+    assert robosuite.__version__ == '1.4.0'
+    
+    copyfile(args.dataset, args.output_dataset)
+    
+    f = h5py.File(args.output_dataset, "r+")
+
+    env_args = json.loads(f["data"].attrs["env_args"])
+    env_name = env_args["env_name"]
+
+    for demo_key in list(f["data"].keys()):
+        ep_data_grp = f["data/{}".format(demo_key)]
+        model_file = ep_data_grp.attrs["model_file"]
+        
+        coverted_model_file = convert_xml(model_file, env_name)
+        ep_data_grp.attrs["model_file"] = coverted_model_file
+        
+    env_args = json.loads(f["data"].attrs["env_args"])
+    env_args["env_version"] = robosuite.__version__ 
+    f["data"].attrs["env_args"] = json.dumps(env_args, indent=4)
+
+    f.close()
